fix(playground): move tool call setting to bottom
