+++ sites/playground/web/src/components/tab-components/mcpTools.vue
@@ -1,9 +1,20 @@
 <script setup>
-import { ref } from 'vue';
-import { TinyCollapse } from '@opentiny/vue';
+import { ref, inject } from 'vue';
+import { TinyCollapse, TinyCheckbox } from '@opentiny/vue';
 import { AgentPanel, McpServerPanel, SkillPanel } from './tool-panels';
 
+const playgroundContext = inject('playgroundContext');
+const { chatConfig } = playgroundContext;
+
 const activePanels = ref(['mcp', 'agent', 'skills']);
+
+const updateAddToolCallContext = (value) => {
+  chatConfig.addToolCallContext = value;
+};
+
+const updateShowThinkingResult = (value) => {
+  chatConfig.showThinkingResult = value;
+};
 </script>
 
 <template>
@@ -13,10 +24,32 @@ const activePanels = ref(['mcp', 'agent', 'skills']);
       <AgentPanel />
       <SkillPanel />
     </tiny-collapse>
+    <div class="tool-call-options">
+      <tiny-checkbox :model-value="chatConfig.addToolCallContext" @update:model-value="updateAddToolCallContext">
+        调用结果添加到上下文
+      </tiny-checkbox>
+    </div>
+    <div class="tool-call-options">
+      <tiny-checkbox :model-value="chatConfig.showThinkingResult" @update:model-value="updateShowThinkingResult">
+        调用结果展示在界面中
+      </tiny-checkbox>
+    </div>
   </div>
 </template>
 
 <style scoped lang="less">
+.tool-call-options {
+  margin-top: 12px;
+  display: flex;
+  align-items: center;
+  font-size: 14px;
+  color: #595959;
+
+  &:last-child {
+    margin-bottom: 16px;
+  }
+}
+
 :deep(.tiny-collapse) {
   border-bottom: none;
